Initial implementation of offset constants generation.
Added the `FieldOffset` type which represents the offset of a field.

Added the `Aligned` and `Unaligned` marker types for whether a field is aligned.

Added the `unsafe_offset_constants` macro to declare the `FieldOffset` for many fields.

Added the `next_field_offset` function and `GetNextFieldOffset` struct as helpers
to calculate the offset of the next field.