Defined some tests.
The tests check that the offsets calculated using the repr_offset crate are the
same as the actual offsets for many cases.

Added some internal features for testing,
including the requirement that the "testing" feature must be enabled in tests.

Defined some internal macros to get the offset of fields in tests from an instance of the type.

Defined the `types_for_tests` module with types used for tests.