Added casting and pointer methods to FieldOffsets.
Added methods for changing the 3 type parameters of FieldOffset individually.

Added `FieldOffset` methods for reading,writing,and replacing the field through a pointer.
For both aligned and unaligned `FieldOffset`s.

Renamed some fields in GetNextFieldOffset

Added a Mem struct with associated constant equivalents of `std::mem::{align_of,size_of}`
(this may not be necessary,it seems that those functions are always evaluated at compile-time)

Changed the min_usize function to not include the panic machinery in debug builds