Added `FieldOffset::add` method, changed "packing" back to "alignment"
Added a `FieldOffset::add` method,to combine `FieldOffsets` to access nested fields.
Implemented `core::ops::Add` with the same semantics.

Replaced all instances of packed with aligned,packing with alignment,
`P` type parameters with `A`, and `Packed` with `Unaligned`.

Moved `Aligned` and `Unaligned` to the alignment module.

Moved `struct_field_offset` module to the root module.

Declared the `Alignment` maker trait, for the `Aligned` and `Unaligned` types.

Declared the `CombinePacking` trait to combine `Aligned` and `Unaligned` on the type level.

Removed the `get_copy_unaligned` method,since it's redundant now.