Added derive macro as dependency,added tests for derive.
Added derive macro as an optional dependency of `repr_offset`,enabled with the "derive" feature.

Added `#[roff(bound = "....")]` attribute to add bounds to the generated impl block.

Added docs to the offsets generated in the derive, mentioning the field that the offset is for.

Added tests for the `ReprOffset` derive macro.

Fixed support for tuple structs, the names of the offsets were wrong.

Added "priv_expensive_test" feature to disable expensive-to-compile tests
(useful to only compile/run tests for local changes).