Added for_examples module,doc example for `FieldOffset::new`.
Added `for_examples` module with types for documentation examples,
so that most examples don't have to construct or derive FieldOffsets to use them.
The items in the module are docuented from Rust 1.41 onwards automatically,
requiring the "for_examples" feature in versions before that one.

Clarified Rust version compatibility in readme,due to some compiler bug that causes
stack overflows and/or memory usage explosion(depending on the version).