Updated readme, more documentation.
Updated the readme to have the same list of features as the `repr_offset` root module.

Updated the `unsafe_struct_field_offsets` example to also use `FieldOffset::replace_mut`.

Added mention of how the safety requirements described in the linked std functions
only apply to the field that the FieldOffset is an offset for.